inventory/backend/requirements.txt
fastapi
uvicorn[standard]
redis<5.0.0
azure-identity
azure-mgmt-network
azure-mgmt-compute
azure-mgmt-appcontainers
azure-mgmt-containerregistry
azure-mgmt-cosmosdb
azure-mgmt-rdbms
azure-mgmt-search
azure-mgmt-cognitiveservices
azure-mgmt-redis
azure-mgmt-containerservice
azure-mgmt-machinelearningservices
azure-mgmt-resource
azure-mgmt-subscription
six
openpyxl